Core/SmartScript: Corrected size of pointId parameter for MOVE_TO_POS SAI action

+++ src/server/game/AI/SmartScripts/SmartScriptMgr.h
@@ -884,7 +884,7 @@ struct SmartAction
 
         struct
         {
-            uint8 pointId;
+            uint32 pointId;
         } MoveToPos;
 
         struct
